+++ src/main/java/com/apptive/marico/service/MemberAuthService.java
@@ -38,19 +38,17 @@ public class MemberAuthService {
     private final RoleRepository roleRepository;
     private final PasswordEncoder passwordEncoder;
 
-    private final TokenProvider tokenProvider;
-    private final AuthenticationManagerBuilder authenticationManagerBuilder;
-    private final RefreshTokenRepository refreshTokenRepository;
     private final VerificationTokenRepository verificationTokenRepository;
+
+    private final CustomUserDetailsService customUserDetailsService;
 
     @Transactional
     public MemberResponseDto signup(MemberRequestDto memberRequestDto) {
+
+        customUserDetailsService.checkEmailAvailability(memberRequestDto.getEmail());
+
         Role userRole = roleRepository.findByName(ROLE_MEMBER).orElseThrow(
                 () -> new CustomException(ROLE_NOT_FOUND));
-
-        if (memberRepository.existsByUserId(memberRequestDto.getEmail())) {
-            throw new CustomException(ALREADY_SAVED_EMAIL);
-        }
 
         Member member = memberRequestDto.toMember(passwordEncoder);
         member.setRoles(Collections.singleton(userRole));
@@ -58,38 +56,6 @@ public class MemberAuthService {
         return MemberResponseDto.toDto(memberRepository.save(member));
     }
 
-    @Transactional
-    public TokenResponseDto login(LoginDto loginDto) {
-        // 1. Login ID/PW 를 기반으로 AuthenticationToken 생성
-        UsernamePasswordAuthenticationToken authenticationToken = loginDto.toAuthentication();
-
-        // 2. 실제로 검증 (사용자 비밀번호 체크) 이 이루어지는 부분
-        //    authenticate 메서드가 실행이 될 때 CustomUserDetailsService 에서 만들었던 loadUserByUsername 메서드가 실행됨
-        Authentication authentication = authenticationManagerBuilder.getObject().authenticate(authenticationToken);
-
-        // 3. 인증 정보를 기반으로 JWT 토큰 생성
-        TokenResponseDto tokenResDto = tokenProvider.generateTokenDto(authentication);
-
-        // 4. RefreshToken 저장
-        RefreshToken refreshToken = RefreshToken.builder()
-                .key(authentication.getName())
-                .value(tokenResDto.getRefreshToken())
-                .build();
-
-        refreshTokenRepository.save(refreshToken);
-
-        // 5. 토큰 발급
-        return tokenResDto;
-    }
-
-    @Transactional
-    public void logout(TokenRequestDto tokenReqDto) {
-        // 로그아웃하려는 사용자의 정보를 가져옴
-        Authentication authentication = tokenProvider.getAuthentication(tokenReqDto.getAccessToken());
-
-        // 저장소에서 해당 사용자의 refresh token 삭제
-        refreshTokenRepository.deleteByKey(authentication.getName());
-    }
 
     public String changePassword(Member member, String password, String code) throws Exception{
         VerificationToken verificationToken = verificationTokenRepository.findByVerificationCode(code);
